+++ sophosfirewall_python/firewallapi.py
@@ -9,379 +9,64 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 permissions and limitations under the License.
 """
 
-import os
-import re
-from ipaddress import IPv4Network, IPv4Address
-import requests
-import xmltodict
 import urllib3
-from jinja2 import Environment, FileSystemLoader
-
+from sophosfirewall_python.api_client import APIClient
+from sophosfirewall_python.firewallrule import FirewallRule
+from sophosfirewall_python.host import (
+    IPHost,
+    IPHostGroup,
+    FQDNHost, 
+    FQDNHostGroup, 
+    URLGroup,
+    IPNetwork,
+    IPRange
+)
+from sophosfirewall_python.service import Service, ServiceGroup
+from sophosfirewall_python.network import Interface, Vlan, Zone
+from sophosfirewall_python.admin import AclRule, Notification
+from sophosfirewall_python.authen import User, AdminAuthen
+from sophosfirewall_python.profile import AdminProfile
+from sophosfirewall_python.ips import IPS
+from sophosfirewall_python.system import Syslog, NotificationList
+from sophosfirewall_python.backup import Backup
+from sophosfirewall_python.reports import Retention
 
 urllib3.disable_warnings()
-
-
-class SophosFirewallIPAddressingError(Exception):
-    """Error raised when invalid IP address detected"""
-
-
-class SophosFirewallAuthFailure(Exception):
-    """Error raised when authentication to firewall fails"""
-
-
-class SophosFirewallAPIError(Exception):
-    """Error raised when an API operation fails"""
-
-
-class SophosFirewallZeroRecords(Exception):
-    """Error raised when a get request returns zero records"""
-
-
-class SophosFirewallOperatorError(Exception):
-    """Error raised when an invalid operator is specified"""
-
-
-class SophosFirewallInvalidArgument(Exception):
-    """Error raised when an invalid argument is specified"""
-
 
 class SophosFirewall:
     """Class used for interacting with the Sophos Firewall XML API"""
 
     def __init__(self, username, password, hostname, port, verify=True):
-        self.username = username
-        self.password = password
-        self.hostname = hostname
-        self.port = port
-        self.url = f"https://{hostname}:{port}/webconsole/APIController"
-        self.verify = verify
-
-    # INTERNAL UTILITY CLASS METHODS
-
-    def _validate_ip_network(self, ip_subnet, mask):
-        """Validate IP network and mask
-
-        Args:
-            ip_subnet (str): IP network address
-            mask (str): Subnet mask
-
-        Raises:
-            SophosFirewallIPAddressingError: Custom error class
-        """
-        try:
-            IPv4Network(f"{ip_subnet}/{mask}")
-        except Exception as exc:
-            raise SophosFirewallIPAddressingError(
-                f"Invalid network or mask provided - {ip_subnet}/{mask}"
-            ) from exc
-
-    def _validate_ip_address(self, ip_address):
-        """Validate IP network and mask
-
-        Args:
-            ip_subnet (str): IP network address
-            mask (str): Subnet mask
-
-        Raises:
-            SophosFirewallIPAddressingError: Custom error class
-        """
-        try:
-            IPv4Address(ip_address)
-        except Exception as exc:
-            raise SophosFirewallIPAddressingError(
-                f"Invalid IP address provided - {ip_address}"
-            ) from exc
-
-    def _post(self, xmldata: str) -> requests.Response:
-        """Post XML request to the firewall returning response as a dict object
-
-        Args:
-            xmldata (str): XML payload
-            verify (bool):  SSL certificate verification. Default=True.
-
-        Returns:
-            requests.Response object
-        """
-        headers = {"Accept": "application/xml"}
-        resp = requests.post(
-            self.url,
-            headers=headers,
-            data={"reqxml": xmldata},
-            verify=self.verify,
-            timeout=30,
+        self.client = APIClient(
+            username=username,
+            password=password,
+            hostname=hostname,
+            port=port,
+            verify=verify
         )
-
-        resp_dict = xmltodict.parse(resp.content.decode())["Response"]
-        if "Status" in resp_dict:
-            if resp_dict["Status"]["@code"] == "534":
-                # IP not allowed in API Access List
-                raise SophosFirewallAPIError(resp_dict["Status"]["#text"])
-
-            if resp_dict["Status"]["@code"] == "532":
-                # API access not enabled
-                raise SophosFirewallAPIError(resp_dict["Status"]["#text"])
-
-        if "Login" in resp_dict:
-            if resp_dict["Login"]["status"] == "Authentication Failure":
-                raise SophosFirewallAuthFailure("Login failed!")
-        return resp
-
-    def _validate_arg(self, arg_name, arg_value, valid_choices):
-        if not arg_value in valid_choices:
-            raise SophosFirewallInvalidArgument(
-                f"Invalid choice for {arg_name} argument, valid choices are {valid_choices}"
-            )
-
-    def submit_template(
-        self,
-        filename: str,
-        template_vars: dict,
-        template_dir: str = None,
-        debug: bool = False,
-    ) -> dict:
-        """Submits XML payload stored as a Jinja2 file
-
-        Args:
-            filename (str): Jinja2 template filename. Place in "templates" directory or configure template_dir.
-            template_vars (dict): Dictionary of variables to inject into the template. Username and password are passed in by default.
-            template_dir (str): Directory to look for templates. Default is "./templates".
-            debug (bool, optional): Enable debug mode to display XML payload. Defaults to False.
-
-        Returns:
-            dict
-        """
-        if not template_dir:
-            template_dir = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)), "templates"
-            )
-        environment = Environment(
-            trim_blocks=True,
-            lstrip_blocks=True,
-            loader=FileSystemLoader(template_dir),
-            autoescape=True,
-        )
-        template = environment.get_template(filename)
-        template_vars["username"] = self.username
-        template_vars["password"] = self.password
-        payload = template.render(**template_vars)
-        if debug:
-            print(f"REQUEST: {payload}")
-        resp = self._post(xmldata=payload)
-
-        resp_dict = xmltodict.parse(resp.content.decode())["Response"]
-        success_pattern = "2[0-9][0-9]"
-        for key in resp_dict:
-            if "Status" in resp_dict[key]:
-                if not re.search(success_pattern, resp_dict[key]["Status"]["@code"]):
-                    raise SophosFirewallAPIError(resp_dict[key])
-        return xmltodict.parse(resp.content.decode())
-
-    def login(self, output_format: str = "dict"):
-        """Test login credentials.
-
-        Args:
-            output_format(str): Output format. Valid options are "dict" or "xml". Defaults to dict.
-        """
-        payload = f"""
-        <Request>
-            <Login>
-                <Username>{self.username}</Username>
-                <Password>{self.password}</Password>
-            </Login>
-        </Request>
-        """
-        resp = self._post(xmldata=payload)
-        if output_format == "xml":
-            return resp.content.decode()
-        return xmltodict.parse(resp.content.decode())
-
-    def get_tag(self, xml_tag: str, output_format: str = "dict"):
-        """Execute a get for a specified XML tag.
-
-        Args:
-            xml_tag (str): XML tag for the request
-            output_format(str): Output format. Valid options are "dict" or "xml". Defaults to dict.
-        """
-        payload = f"""
-        <Request>
-            <Login>
-                <Username>{self.username}</Username>
-                <Password>{self.password}</Password>
-            </Login>
-            <Get>
-                <{xml_tag}>
-                </{xml_tag}>
-            </Get>
-        </Request>
-        """
-        resp = self._post(xmldata=payload)
-        self._error_check(resp, xml_tag)
-        if output_format == "xml":
-            return resp.content.decode()
-        return xmltodict.parse(resp.content.decode())
-
-    def get_tag_with_filter(
-        self,
-        xml_tag: str,
-        key: str,
-        value: str,
-        operator: str = "like",
-        output_format: str = dict,
-    ):
-        """Execute a get for a specified XML tag with filter criteria.
-
-        Args:
-            xml_tag (str): XML tag for the request.
-            key (str): Search key
-            value (str): Search value
-            operator (str, optional): Operator for search (“=”,”!=”,”like”). Defaults to "like".
-            output_format(str): Output format. Valid options are "dict" or "xml". Defaults to dict.
-        """
-        valid_operators = ["=", "!=", "like"]
-        if operator not in valid_operators:
-            raise SophosFirewallOperatorError(
-                f"Invalid operator '{operator}'!  Supported operators: [ {', '.join(valid_operators)} ]"
-            )
-        payload = f"""
-        <Request>
-            <Login>
-                <Username>{self.username}</Username>
-                <Password>{self.password}</Password>
-            </Login>
-            <Get>
-                <{xml_tag}>
-                    <Filter>
-                        <key name="{key}" criteria="{operator}">{value}</key>
-                    </Filter>
-                </{xml_tag}>
-            </Get>
-        </Request>
-        """
-        resp = self._post(xmldata=payload)
-        self._error_check(resp, xml_tag)
-        if output_format == "xml":
-            return resp.content.decode()
-        return xmltodict.parse(resp.content.decode())
-
-    def remove(self, xml_tag: str, name: str, output_format: str = "dict"):
-        """Remove an object from the firewall.
-
-        Args:
-            xml_tag (str): The XML tag indicating the type of object to be removed.
-            name (str): The name of the object to be removed.
-            output_format (str): Output format. Valid options are "dict" or "xml". Defaults to dict.
-        """
-        payload = f"""
-        <Request>
-            <Login>
-                <Username>{self.username}</Username>
-                <Password>{self.password}</Password>
-            </Login>
-            <Remove>
-              <{xml_tag}>
-                <Name>{name}</Name>
-              </{xml_tag}>
-            </Remove>
-        </Request>
-        """
-        resp = self._post(xmldata=payload)
-        self._error_check(resp, xml_tag)
-        if output_format == "xml":
-            return resp.content.decode()
-        return xmltodict.parse(resp.content.decode())
-
-    def update(
-        self,
-        xml_tag: str,
-        update_params: dict,
-        name: str = None,
-        output_format: str = "dict",
-        debug: bool = False,
-    ):
-        """Update an existing object on the firewall.
-
-        Args:
-            xml_tag (str): The XML tag indicating the type of object to be updated.
-            update_params (dict): Keys/values to be updated. Keys must match an existing XML key.
-            name (str, optional): The name of the object to be updated, if applicable.
-            output_format(str): Output format. Valid options are "dict" or "xml". Defaults to dict.
-            debug (bool): Displays the XML payload that was submitted
-        """
-        if name:
-            resp = self.get_tag_with_filter(
-                xml_tag=xml_tag, key="Name", value=name, operator="="
-            )
-        else:
-            resp = self.get_tag(xml_tag=xml_tag)
-
-        for key in update_params:
-            resp["Response"][xml_tag][key] = update_params[key]
-
-        update_body = {}
-        update_body[xml_tag] = resp["Response"][xml_tag]
-        xml_update_body = xmltodict.unparse(update_body, pretty=True).lstrip(
-            '<?xml version="1.0" encoding="utf-8"?>'
-        )
-        payload = f"""
-        <Request>
-            <Login>
-                <Username>{self.username}</Username>
-                <Password>{self.password}</Password>
-            </Login>
-            <Set operation="update"> 
-                {xml_update_body}
-            </Set>
-        </Request>
-        """
-        if debug:
-            print(payload)
-        resp = self._post(xmldata=payload)
-        self._error_check(resp, xml_tag)
-        if output_format == "xml":
-            return resp.content.decode()
-        return xmltodict.parse(resp.content.decode())
-
-    def _dict_to_lower(self, target_dict):
-        """Convert the keys of a dictionary to lower-case
-
-        Args:
-            target_dict (dict): Dictionary to be converted
-
-        Returns:
-            dict: Dictionary with all keys converted to lower case
-        """
-        return {key.lower(): val for key, val in target_dict.items()}
-
-    def _error_check(self, api_response, xml_tag):
-        """Check for errors in the API response and raise exception if present
-
-        Args:
-            api_response (Requests.response): The response object returned from the requests module
-            xml_tag (str): The XML tag being operated on
-
-        Raises:
-            SophosFirewallZeroRecords: Error raised when there are no records matching the request parameters
-            SophosFirewallAPIError: Error raised when there is a problem with the request parameters
-        """
-        response = xmltodict.parse(api_response.content.decode())["Response"]
-        lower_response = self._dict_to_lower(response)
-        if xml_tag.lower() in lower_response:
-            resp_dict = lower_response[xml_tag.lower()]
-            if "Status" in resp_dict:
-                if (
-                    resp_dict["Status"] == "Number of records Zero."
-                    or resp_dict["Status"] == "No. of records Zero."
-                ):
-                    raise SophosFirewallZeroRecords(resp_dict["Status"])
-                if "@code" in resp_dict["Status"]:
-                    if not resp_dict["Status"]["@code"].startswith("2"):
-                        raise SophosFirewallAPIError(
-                            f"{resp_dict['Status']['@code']}: {resp_dict['Status']['#text']}"
-                        )
-        else:
-            raise SophosFirewallAPIError(
-                str(xmltodict.parse(api_response.content.decode()))
-            )
+        self.firewall_rule = FirewallRule(self.client)
+        self.ip_host = IPHost(self.client)
+        self.ip_hostgroup = IPHostGroup(self.client)
+        self.fqdn_host = FQDNHost(self.client)
+        self.fqdn_hostgroup = FQDNHostGroup(self.client)
+        self.ip_network = IPNetwork(self.client)
+        self.ip_range = IPRange(self.client)
+        self.service = Service(self.client)
+        self.service_group = ServiceGroup(self.client)
+        self.interface = Interface(self.client)
+        self.vlan = Vlan(self.client)
+        self.acl_rule = AclRule(self.client)
+        self.user = User(self.client)
+        self.admin_profile = AdminProfile(self.client)
+        self.admin_auth = AdminAuthen(self.client)
+        self.zone = Zone(self.client)
+        self.ips = IPS(self.client)
+        self.syslog_server = Syslog(self.client)
+        self.notification = Notification(self.client)
+        self.notification_list = NotificationList(self.client)
+        self.backup = Backup(self.client)
+        self.report_retention = Retention(self.client)
+        self.url_group = URLGroup(self.client)
 
     # METHODS FOR OBJECT RETRIEVAL (GET)
 
@@ -391,13 +76,8 @@ class SophosFirewall:
         Args:
             name (str, optional): Firewall Rule name.  Returns all rules if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
-            debug(bool, optional): Enable debug mode
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="FirewallRule", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="FirewallRule")
+        return self.firewall_rule.get(name=name, operator=operator)
 
     def get_ip_host(
         self, name: str = None, ip_address: str = None, operator: str = "="
@@ -409,19 +89,40 @@ class SophosFirewall:
             ip_address (str, optional): Query by IP Address.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="IPHost", key="Name", value=name, operator=operator
-            )
-        if ip_address:
-            return self.get_tag_with_filter(
-                xml_tag="IPHost",
-                key="IPAddress",
-                value=ip_address,
-                operator=operator,
-            )
-        return self.get_tag(xml_tag="IPHost")
+        return self.ip_host.get(name, ip_address, operator)
     
+    def get_ip_hostgroup(self, name: str = None, operator: str = "="):
+        """Get IP hostgroup(s)
+
+        Args:
+            name (str, optional): Name of IP host group. Returns all if not specified.
+            operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
+        """
+        return self.ip_hostgroup.get(name, operator)
+    
+    def get_ip_network(
+        self, name: str = None, ip_address: str = None, operator: str = "="
+    ):
+        """Get IP Network object(s)
+
+        Args:
+            name (str, optional): IP object name. Returns all objects if not specified.
+            ip_address (str, optional): Query by IP Address.
+            operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
+        """
+        return self.ip_network.get(name, ip_address, operator)
+    
+    def get_ip_range(
+        self, name: str = None, operator: str = "="
+    ):
+        """Get IP Range object(s)
+
+        Args:
+            name (str, optional): IP object name. Returns all objects if not specified.
+            operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
+        """
+        return self.ip_range.get(name, operator)
+
     def get_fqdn_host(
         self, name: str = None, operator: str = "="
     ):
@@ -431,12 +132,7 @@ class SophosFirewall:
             name (str, optional): FQDN Host name. Returns all objects if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="FQDNHost", key="Name", value=name, operator=operator
-            )
-        
-        return self.get_tag(xml_tag="FQDNHost")
+        return self.fqdn_host.get(name, operator)
     
     def get_fqdn_hostgroup(
         self, name: str = None, operator: str = "="
@@ -447,12 +143,8 @@ class SophosFirewall:
             name (str, optional): FQDN HostGroup name. Returns all objects if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="FQDNHostGroup", key="Name", value=name, operator=operator
-            )
-        
-        return self.get_tag(xml_tag="FQDNHostGroup")
+        return self.fqdn_hostgroup.get(name, operator)
+
     
     def get_service_group(
         self, name: str = None, operator: str = "="
@@ -463,12 +155,7 @@ class SophosFirewall:
             name (str, optional): Service Group name. Returns all objects if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="ServiceGroup", key="Name", value=name, operator=operator
-            )
-        
-        return self.get_tag(xml_tag="ServiceGroup")
+        return self.service_group.get(name, operator)
 
     def get_interface(self, name: str = None, operator: str = "="):
         """Get Interface object(s)
@@ -477,11 +164,8 @@ class SophosFirewall:
             name (str, optional): Interface name. Returns all objects if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="Interface", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="Interface")
+        return self.interface.get(name, operator)
+
 
     def get_vlan(self, name: str = None, operator: str = "="):
         """Get VLAN object(s)
@@ -490,27 +174,7 @@ class SophosFirewall:
             name (str, optional): VLAN name. Returns all objects if not specified.
             operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="VLAN", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="VLAN")
-
-    def get_ip_hostgroup(self, name: str = None, operator: str = "="):
-        """Get IP hostgroup(s)
-
-        Args:
-            name (str, optional): Name of IP host group. Returns all if not specified.
-            operator (str, optional): Operator for search. Default is "=". Valid operators: =, !=, like.
-        """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="IPHostGroup",
-                key="Name",
-                value=name,
-                operator=operator,
-            )
-        return self.get_tag(xml_tag="IPHostGroup")
+        return self.vlan.get(name, operator)
 
     def get_acl_rule(self, name: str = None, operator: str = "="):
         """Get ACL rules
@@ -522,11 +186,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="LocalServiceACL", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="LocalServiceACL")
+        return self.acl_rule.get(name, operator)
+
 
     def get_user(self, name: str = None, operator: str = "="):
         """Get local users
@@ -538,11 +199,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="User", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="User")
+        return self.user.get(name, operator)
+
 
     def get_admin_profile(self, name: str = None, operator: str = "="):
         """Get admin profiles
@@ -554,14 +212,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="AdministrationProfile",
-                key="Name",
-                value=name,
-                operator=operator,
-            )
-        return self.get_tag(xml_tag="AdministrationProfile")
+        return self.admin_profile.get(name, operator)
+
 
     def get_zone(self, name: str = None, operator: str = "="):
         """Get zone(s)
@@ -573,11 +225,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="Zone", key="Name", value=name, operator=operator
-            )
-        return self.get_tag(xml_tag="Zone")
+        return self.zone.get(name, operator)
 
     def get_admin_authen(self):
         """Get admin authentication settings
@@ -585,7 +233,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        return self.get_tag(xml_tag="AdminAuthentication")
+        return self.admin_auth.get()
 
     def get_ips_policy(self, name: str = None):
         """Get IPS policy
@@ -596,9 +244,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(xml_tag="IPSPolicy", key="Name", value=name)
-        return self.get_tag(xml_tag="IPSPolicy")
+        return self.ips.get(name)
 
     def get_syslog_server(self, name: str = None):
         """Get syslog server.
@@ -609,11 +255,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="SyslogServers", key="Name", value=name
-            )
-        return self.get_tag(xml_tag="SyslogServers")
+        return self.syslog_server.get(name)
 
     def get_notification(self, name: str = None):
         """Get notification.
@@ -624,11 +266,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="Notification", key="Name", value=name
-            )
-        return self.get_tag(xml_tag="Notification")
+        return self.notification.get(name)
+
 
     def get_notification_list(self, name: str = None):
         """Get notification list.
@@ -639,11 +278,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="NotificationList", key="Name", value=name
-            )
-        return self.get_tag(xml_tag="NotificationList")
+        self.notification_list.get(name)
 
     def get_backup(self, name: str = None):
         """Get backup details.
@@ -654,11 +289,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="BackupRestore", key="Name", value=name
-            )
-        return self.get_tag(xml_tag="BackupRestore")
+        return self.backup.get(name)
 
     def get_reports_retention(self, name: str = None):
         """Get Reports retention period.
@@ -669,11 +300,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="DataManagement", key="Name", value=name
-            )
-        return self.get_tag(xml_tag="DataManagement")
+        return self.report_retention.get(name)
 
     def get_admin_settings(self):
         """Get Web Admin Settings (Administration > Settings)
@@ -681,7 +308,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        return self.get_tag(xml_tag="AdminSettings")
+        return self.client.get_tag(xml_tag="AdminSettings")
 
     def get_dns_forwarders(self):
         """Get DNS forwarders.
@@ -689,7 +316,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        return self.get_tag(xml_tag="DNS")
+        return self.client.get_tag(xml_tag="DNS")
 
     def get_snmpv3_user(self):
         """Get SNMP v3 Users
@@ -697,7 +324,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        return self.get_tag(xml_tag="SNMPv3User")
+        return self.client.get_tag(xml_tag="SNMPv3User")
 
     def get_urlgroup(self, name: str = None, operator: str = "="):
         """Get URLGroup(s)
@@ -709,11 +336,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="WebFilterURLGroup", key="Name", operator=operator, value=name
-            )
-        return self.get_tag(xml_tag="WebFilterURLGroup")
+        return self.url_group.get(name, operator)
+
 
     def get_service(
         self,
@@ -733,36 +357,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if name:
-            return self.get_tag_with_filter(
-                xml_tag="Services", key="Name", value=name, operator=operator
-            )
-        if dst_proto and dst_port:
-            resp = self.get_tag(xml_tag="Services")
-            svc_list = resp["Response"]["Services"].copy()
-            for svc in svc_list:
-                matched = False
-                if isinstance(svc["ServiceDetails"]["ServiceDetail"], dict):
-                    port = svc["ServiceDetails"]["ServiceDetail"].get("DestinationPort")
-                    if "Protocol" in svc["ServiceDetails"]["ServiceDetail"]:
-                        proto = svc["ServiceDetails"]["ServiceDetail"]["Protocol"]
-                    if "ProtocolName" in svc["ServiceDetails"]["ServiceDetail"]:
-                        proto = svc["ServiceDetails"]["ServiceDetail"]["ProtocolName"]
-                    if proto == dst_proto.upper() and port == dst_port:
-                        matched = True
-                elif isinstance(svc["ServiceDetails"]["ServiceDetail"], list):
-                    for subsvc in svc["ServiceDetails"]["ServiceDetail"]:
-                        port = subsvc.get("DestinationPort")
-                        if "Protocol" in subsvc:
-                            proto = subsvc["Protocol"]
-                        if "ProtocolName" in subsvc:
-                            proto = subsvc["ProtocolName"]
-                        if proto == dst_proto.upper() and port == dst_port:
-                            matched = True
-                if not matched:
-                    resp["Response"]["Services"].remove(svc)
-            return resp
-        return self.get_tag(xml_tag="Services")
+        return self.service.get(name, operator, dst_proto, dst_port)
 
     # METHODS FOR OBJECT CREATION
 
@@ -786,10 +381,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        resp = self.submit_template(
-            "createfwrule.j2", template_vars=rule_params, debug=debug
-        )
-        return resp
+        return self.firewall_rule.create(rule_params, debug)      
 
     def create_ip_network(
         self,
@@ -808,13 +400,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        self._validate_ip_network(ip_network, mask)
-
-        params = {"name": name, "ip_network": ip_network, "mask": mask}
-        resp = self.submit_template(
-            "createipnetwork.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.ip_network.create(name, ip_network, mask, debug)
 
     def create_ip_host(self, name: str, ip_address: str, debug: bool = False):
         """Create IP address object
@@ -826,13 +412,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        self._validate_ip_address(ip_address)
-
-        params = {"name": name, "ip_address": ip_address}
-        resp = self.submit_template(
-            "createiphost.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.ip_host.create(name, ip_address, debug)
     
     def create_fqdn_host(self, name: str,
                          fqdn: str,
@@ -850,12 +430,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary.
         """
-
-        params = {"name": name, "description": description, "fqdn": fqdn, "fqdn_group_list": fqdn_group_list}
-        resp = self.submit_template(
-            "createfqdnhost.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.fqdn_host.create(name, fqdn, fqdn_group_list, description, debug)
 
     def create_fqdn_hostgroup(self, name: str,
                          fqdn_host_list: list = None,
@@ -871,12 +446,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary.
         """
-
-        params = {"name": name, "description": description, "fqdn_host_list": fqdn_host_list}
-        resp = self.submit_template(
-            "createfqdnhostgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        self.fqdn_hostgroup.create(name, fqdn_host_list, description, debug)
+        
 
     def create_ip_range(
         self,
@@ -895,14 +466,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        self._validate_ip_address(start_ip)
-        self._validate_ip_address(end_ip)
-
-        params = {"name": name, "start_ip": start_ip, "end_ip": end_ip}
-        resp = self.submit_template(
-            "createiprange.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.ip_range.create(name, start_ip, end_ip, debug)
 
     def create_service(
         self,
@@ -923,11 +487,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        params = {"name": name, "service_list": service_list, "type": service_type}
-        resp = self.submit_template(
-            "createservice.j2", template_vars=params, debug=debug
-        )
-        return resp
+        self.service.create(name, service_type, service_list, debug)
+        
 
     def create_service_group(self, name: str,
                          service_list: list = None,
@@ -943,12 +504,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary.
         """
-
-        params = {"name": name, "description": description, "service_list": service_list}
-        resp = self.submit_template(
-            "createservicegroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.service_group.create(name, service_list, description, debug)
 
     def create_ip_hostgroup(
         self,
@@ -967,11 +523,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        params = {"name": name, "description": description, "host_list": host_list}
-        resp = self.submit_template(
-            "createiphostgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        self.ip_hostgroup.create(name, host_list, description, debug)
+        
 
     def create_urlgroup(self, name: str, domain_list: list, debug: bool = False):
         """Create a web URL Group
@@ -984,11 +537,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        params = {"name": name, "domain_list": domain_list}
-        resp = self.submit_template(
-            "createurlgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.url_group.create(name, domain_list, debug)
+        
 
     def create_user(self, debug: bool = False, **kwargs):
         """Create a User
@@ -1023,9 +573,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
+        return self.user.create(debug, **kwargs)
 
-        resp = self.submit_template("createuser.j2", template_vars=kwargs, debug=debug)
-        return resp
 
     def update_user_password(
         self, username: str, new_password: str, debug: bool = False
@@ -1040,17 +589,8 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        # Get the existing user
-        resp = self.get_user(name=username)
-        user_params = resp["Response"]["User"]
-        user_params["Password"] = new_password
-        user_params.pop("PasswordHash")
-
-        # Update the user
-        resp = self.submit_template(
-            "updateuserpassword.j2", template_vars=user_params, debug=debug
-        )
-        return resp
+        return self.user.update_user_password(username, new_password, debug)
+        
 
     def update_admin_password(
         self, current_password: str, new_password: str, debug: bool = False
@@ -1065,12 +605,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        params = {"current_password": current_password, "new_password": new_password}
-
-        resp = self.submit_template(
-            "updateadminpassword.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.user.update_admin_password(current_password, new_password, debug)
 
     def update_urlgroup(
         self, name: str, domain_list: list, action: str = "add", debug: bool = False
@@ -1086,48 +621,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if not isinstance(domain_list, list):
-            raise SophosFirewallInvalidArgument(
-                "The update_urlgroup() argument `domain_list` must be of type list!"
-            )
-
-        if action:
-            self._validate_arg(
-                arg_name="action",
-                arg_value=action,
-                valid_choices=["add", "remove", "replace"],
-            )
-
-        # Get the existing URL list first, if any
-        resp = self.get_urlgroup(name=name)
-        if "URLlist" in resp["Response"]["WebFilterURLGroup"]:
-            exist_list = (
-                resp.get("Response").get("WebFilterURLGroup").get("URLlist").get("URL")
-            )
-        else:
-            exist_list = None
-        if action == "replace":
-            exist_list = None
-        new_domain_list = []
-        if exist_list:
-            if isinstance(exist_list, str):
-                new_domain_list.append(exist_list)
-            elif isinstance(exist_list, list):
-                for domain in exist_list:
-                    new_domain_list.append(domain)
-        for domain in domain_list:
-            if action.lower() == "add" and domain not in new_domain_list:
-                new_domain_list.append(domain)
-            elif action.lower() == "remove" and domain in new_domain_list:
-                new_domain_list.remove(domain)
-            elif action.lower() == "replace":
-                new_domain_list.append(domain)
-
-        params = {"name": name, "domain_list": new_domain_list}
-        resp = self.submit_template(
-            "updateurlgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.url_group.update(name, domain_list, action, debug)
 
     def update_service(
         self,
@@ -1151,110 +645,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        if not isinstance(service_list, list):
-            raise SophosFirewallInvalidArgument(
-                "The update_service() argument `service_list` must be of type list!"
-            )
-
-        if action:
-            self._validate_arg(
-                arg_name="action",
-                arg_value=action,
-                valid_choices=["add", "remove", "replace"],
-            )
-
-        # Get the existing Service list first
-        resp = self.get_service(name=name)
-        if "ServiceDetail" in resp["Response"]["Services"]["ServiceDetails"]:
-            exist_list = (
-                resp.get("Response")
-                .get("Services")
-                .get("ServiceDetails")
-                .get("ServiceDetail")
-            )
-        else:
-            exist_list = None
-
-        # Add src_port to input if not present
-        for service in service_list:
-            if not "src_port" in service:
-                service["src_port"] = "1:65535"
-        if action == "replace":
-            exist_list = None
-        new_service_list = []
-        if exist_list:
-            if isinstance(exist_list, dict):
-                if service_type == "TCPorUDP":
-                    new_service_list.append(
-                        {
-                            "src_port": exist_list["SourcePort"],
-                            "dst_port": exist_list["DestinationPort"],
-                            "protocol": exist_list["Protocol"],
-                        }
-                    )
-                if service_type == "IP":
-                    new_service_list.append(
-                        {
-                            "protocol": exist_list["ProtocolName"]
-                        }
-                    )
-                if service_type == "ICMP":
-                    new_service_list.append(
-                        {
-                            "icmp_type": exist_list["ICMPType"],
-                            "icmp_code": exist_list["ICMPCode"]
-                        }
-                    )
-                if service_type == "ICMPv6":
-                    new_service_list.append(
-                        {
-                            "icmp_type": exist_list["ICMPv6Type"],
-                            "icmp_code": exist_list["ICMPv6Code"]
-                        }
-                    )
-            elif isinstance(exist_list, list):
-                for service in exist_list:
-                    if service_type == "TCPorUDP":
-                        new_service_list.append(
-                            {
-                                "src_port": service["SourcePort"],
-                                "dst_port": service["DestinationPort"],
-                                "protocol": service["Protocol"],
-                            }
-                        )
-                    if service_type == "IP":
-                        new_service_list.append(
-                            {
-                                "protocol": service["ProtocolName"]
-                            }
-                        )
-                    if service_type == "ICMP":
-                        new_service_list.append(
-                            {
-                                "icmp_type": service["ICMPType"],
-                                "icmp_code": service["ICMPCode"]
-                            }
-                        )
-                    if service_type == "ICMPv6":
-                        new_service_list.append(
-                            {
-                                "icmp_type": service["ICMPv6Type"],
-                                "icmp_code": service["ICMPv6Code"]
-                            }
-                        )
-        for service in service_list:
-            if action.lower() == "add" and service not in new_service_list:
-                new_service_list.append(service)
-            elif action.lower() == "remove" and service in new_service_list:
-                new_service_list.remove(service)
-            elif action.lower() == "replace":
-                new_service_list.append(service)
-
-        params = {"name": name, "service_list": new_service_list, "type": service_type}
-        resp = self.submit_template(
-            "updateservice.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.service.update(name, service_type, service_list, action, debug)
 
     def update_ip_hostgroup(
         self,
@@ -1276,48 +667,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        # Get the existing Host list first, if any
-
-        if action:
-            self._validate_arg(
-                arg_name="action",
-                arg_value=action,
-                valid_choices=["add", "remove", "replace"],
-            )
-
-        resp = self.get_ip_hostgroup(name=name)
-        if "HostList" in resp["Response"]["IPHostGroup"]:
-            exist_list = (
-                resp.get("Response").get("IPHostGroup").get("HostList").get("Host")
-            )
-        else:
-            exist_list = None
-
-        if action.lower() == "replace":
-            exist_list = None
-
-        new_host_list = []
-        if exist_list:
-            if isinstance(exist_list, str):
-                new_host_list.append(exist_list)
-            elif isinstance(exist_list, list):
-                new_host_list = exist_list
-        for ip_host in host_list:
-            if action:
-                if action.lower() == "add" and not ip_host in new_host_list:
-                    new_host_list.append(ip_host)
-                elif action.lower() == "remove" and ip_host in new_host_list:
-                    new_host_list.remove(ip_host)
-                elif action.lower() == "replace":
-                    new_host_list.append(ip_host)
-        if not description:
-            description = resp.get("Response").get("IPHostGroup").get("Description")
-
-        params = {"name": name, "description": description, "host_list": new_host_list}
-        resp = self.submit_template(
-            "updateiphostgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.ip_hostgroup.update(name, host_list, description, action, debug)
 
     def update_fqdn_hostgroup(
         self,
@@ -1339,48 +689,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        # Get the existing Host list first, if any
-
-        if action:
-            self._validate_arg(
-                arg_name="action",
-                arg_value=action,
-                valid_choices=["add", "remove", "replace"],
-            )
-
-        resp = self.get_fqdn_hostgroup(name=name)
-        if "FQDNHostList" in resp["Response"]["FQDNHostGroup"]:
-            exist_list = (
-                resp.get("Response").get("FQDNHostGroup").get("FQDNHostList").get("FQDNHost")
-            )
-        else:
-            exist_list = None
-
-        if action.lower() == "replace":
-            exist_list = None
-
-        new_host_list = []
-        if exist_list:
-            if isinstance(exist_list, str):
-                new_host_list.append(exist_list)
-            elif isinstance(exist_list, list):
-                new_host_list = exist_list
-        for fqdn_host in fqdn_host_list:
-            if action:
-                if action.lower() == "add" and not fqdn_host in new_host_list:
-                    new_host_list.append(fqdn_host)
-                elif action.lower() == "remove" and fqdn_host in new_host_list:
-                    new_host_list.remove(fqdn_host)
-                elif action.lower() == "replace":
-                    new_host_list.append(fqdn_host)
-        if not description:
-            description = resp.get("Response").get("FQDNHostGroup").get("Description")
-
-        params = {"name": name, "description": description, "fqdn_host_list": new_host_list}
-        resp = self.submit_template(
-            "updatefqdnhostgroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.fqdn_hostgroup.update(name, description,fqdn_host_list, action, debug)
 
     def update_service_group(
         self,
@@ -1403,47 +712,7 @@ class SophosFirewall:
             dict: XML response converted to Python dictionary
         """
         # Get the existing Host list first, if any
-
-        if action:
-            self._validate_arg(
-                arg_name="action",
-                arg_value=action,
-                valid_choices=["add", "remove", "replace"],
-            )
-
-        resp = self.get_service_group(name=name)
-        if "ServiceList" in resp["Response"]["ServiceGroup"]:
-            exist_list = (
-                resp.get("Response").get("ServiceGroup").get("ServiceList").get("Service")
-            )
-        else:
-            exist_list = None
-
-        if action.lower() == "replace":
-            exist_list = None
-
-        new_service_list = []
-        if exist_list:
-            if isinstance(exist_list, str):
-                new_service_list.append(exist_list)
-            elif isinstance(exist_list, list):
-                new_service_list = exist_list
-        for service_name in service_list:
-            if action:
-                if action.lower() == "add" and not service_name in new_service_list:
-                    new_service_list.append(service_name)
-                elif action.lower() == "remove" and service_name in new_service_list:
-                    new_service_list.remove(service_name)
-                elif action.lower() == "replace":
-                    new_service_list.append(service_name)
-        if not description:
-            description = resp.get("Response").get("ServiceGroup").get("Description")
-
-        params = {"name": name, "description": description, "service_list": new_service_list}
-        resp = self.submit_template(
-            "updateservicegroup.j2", template_vars=params, debug=debug
-        )
-        return resp
+        return self.service_group.update(name, service_list, description, action, debug)
 
     def update_backup(self, backup_params: dict, debug: bool = False):
         """Updates scheduled backup settings
@@ -1470,20 +739,7 @@ class SophosFirewall:
         Returns:
             dict: XML response converted to Python dictionary
         """
-        updated_params = {}
-        current_params = self.get_backup()["Response"]["BackupRestore"][
-            "ScheduleBackup"
-        ]
-        for param in current_params:
-            if param in backup_params:
-                updated_params[param] = backup_params[param]
-            else:
-                updated_params[param] = current_params[param]
-
-        resp = self.submit_template(
-            "updatebackup.j2", template_vars=updated_params, debug=debug
-        )
-        return resp
+        return self.backup.update(backup_params, debug)
 
     def update_service_acl(
         self,
@@ -1501,33 +757,4 @@ class SophosFirewall:
             verify (bool, optional): SSL Certificate checking. Defaults to True.
             debug (bool, optional): Enable debug mode. Defaults to False.
         """
-        if action:
-            self._validate_arg(
-                arg_name="action", arg_value=action, valid_choices=["add", "remove"]
-            )
-        resp = self.get_acl_rule()
-
-        exist_hosts = resp["Response"]["LocalServiceACL"]["Hosts"]["Host"]
-        exist_services = resp["Response"]["LocalServiceACL"]["Services"]["Service"]
-
-        if not host_list:
-            host_list = []
-        if not service_list:
-            service_list = []
-
-        if action == "add":
-            template_vars = {
-                "host_list": exist_hosts + host_list,
-                "service_list": exist_services + service_list,
-            }
-        elif action == "remove":
-            for host in host_list:
-                exist_hosts.remove(host)
-            for service in service_list:
-                exist_services.remove(service)
-            template_vars = {"host_list": exist_hosts, "service_list": exist_services}
-        resp = self.submit_template(
-            "updateserviceacl.j2", template_vars=template_vars, debug=debug
-        )
-
-        return resp
+        return self.acl_rule.update(host_list, service_list, action, debug)
